+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+  - Removed sources of excessive cpu usage
+
 ## [0.9.0] - 2018-07-23
 
 ### Added