+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.9.2] - 2018-09-04
+
 ### Fixed
 
   - Resolve panic when statefile contained bad data. State is now initialized
@@ -255,7 +257,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     stop
   - Ability to upload logs by streaming them over STDIN
 
-[Unreleased]: https://github.com/timberio/agent/compare/v0.9.1...HEAD
+[Unreleased]: https://github.com/timberio/agent/compare/v0.9.2...HEAD
+[0.9.2]: https://github.com/timberio/agent/compare/v0.9.1...v0.9.2
 [0.9.1]: https://github.com/timberio/agent/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/timberio/agent/compare/v0.8.3...v0.9.0
 [0.8.3]: https://github.com/timberio/agent/compare/v0.8.2...v0.8.3